requirements.txt
websockets>=11.0.0
Pillow>=10.0.0
